Move queries into children
